Refactor: Consolidate sale and lot timer helper and make sure days are counted the same in app and web
