+++ src/v2/Apps/Auction/Components/AuctionDetails/SaleDetailTimer.tsx
@@ -3,6 +3,7 @@ import { createFragmentContainer, graphql } from "react-relay"
 import * as React from "react"
 import { Flex, Text } from "@artsy/palette"
 import { useTimer } from "v2/Utils/Hooks/useTimer"
+import { getSaleOrLotTimerInfo } from "v2/Utils/getSaleOrLotTimerInfo"
 
 export interface SaleDetailTimerProps {
   sale: SaleDetailTimer_sale
@@ -18,7 +19,7 @@ export const SaleDetailTimer: React.FC<SaleDetailTimerProps> = ({ sale }) => {
     return null
   }
 
-  const timerCopy = getTimerCopy(time, hasStarted, hasEnded)
+  const timerCopy = getSaleOrLotTimerInfo(time, hasStarted, hasEnded, true)
 
   return (
     <Flex alignItems="center" flexDirection="column">
@@ -42,56 +43,56 @@ export const SaleDetailTimerFragmentContainer = createFragmentContainer(
   }
 )
 
-export const getTimerCopy = (time, hasStarted, lotsAreClosing) => {
-  const { days, hours, minutes, seconds } = time
+// export const getTimerCopy = (time, hasStarted, lotsAreClosing) => {
+//   const { days, hours, minutes, seconds } = time
 
-  const parsedDays = parseInt(days, 10)
-  const parsedHours = parseInt(hours, 10)
-  const parsedMinutes = parseInt(minutes, 10)
-  const parsedSeconds = parseInt(seconds, 10)
+//   const parsedDays = parseInt(days, 10)
+//   const parsedHours = parseInt(hours, 10)
+//   const parsedMinutes = parseInt(minutes, 10)
+//   const parsedSeconds = parseInt(seconds, 10)
 
-  let copy = ""
-  let color = "blue100"
+//   let copy = ""
+//   let color = "blue100"
 
-  // Sale has not yet started
-  if (!hasStarted) {
-    copy = relativeDaysUntil(parsedDays, "Bidding Starts")
-  }
-  // Lots are closing
-  else if (lotsAreClosing) {
-    copy = "Lots are closing"
-    color = "red100"
-  }
-  // Lots haven't yet closed
-  else {
-    // 1 hour or less
-    if (parsedDays < 1 && parsedHours < 1) {
-      copy = `${parsedMinutes}m ${parsedSeconds}s`
-      color = "red100"
-    }
-    // More than 24 hours
-    else if (parsedDays >= 1) {
-      copy = `${parsedDays + 1} Day${parsedDays >= 1 ? "s" : ""}`
-    }
+//   // Sale has not yet started
+//   if (!hasStarted) {
+//     copy = relativeDaysUntil(parsedDays, "Bidding Starts")
+//   }
+//   // Lots are closing
+//   else if (lotsAreClosing) {
+//     copy = "Lots are closing"
+//     color = "red100"
+//   }
+//   // Lots haven't yet closed
+//   else {
+//     // 1 hour or less
+//     if (parsedDays < 1 && parsedHours < 1) {
+//       copy = `${parsedMinutes}m ${parsedSeconds}s`
+//       color = "red100"
+//     }
+//     // More than 24 hours
+//     else if (parsedDays >= 1) {
+//       copy = `${parsedDays + 1} Day${parsedDays >= 1 ? "s" : ""}`
+//     }
 
-    // 1-24 hours until close
-    else if (parsedDays < 1 && parsedHours >= 1) {
-      copy = `${parsedHours}h ${parsedMinutes}m`
-      color = "red100"
-    }
+//     // 1-24 hours until close
+//     else if (parsedDays < 1 && parsedHours >= 1) {
+//       copy = `${parsedHours}h ${parsedMinutes}m`
+//       color = "red100"
+//     }
 
-    copy += " Until Lots Start Closing"
-  }
+//     copy += " Until Lots Start Closing"
+//   }
 
-  return { copy, color }
-}
+//   return { copy, color }
+// }
 
-function relativeDaysUntil(parsedDays, suffix) {
-  if (parsedDays === 0) {
-    return `${suffix} Today`
-  } else if (parsedDays === 1) {
-    return `${parsedDays} Day Until ${suffix}`
-  } else {
-    return `${parsedDays} Days Until ${suffix}`
-  }
-}
+// function relativeDaysUntil(parsedDays, suffix) {
+//   if (parsedDays === 0) {
+//     return `${suffix} Today`
+//   } else if (parsedDays === 1) {
+//     return `${parsedDays} Day Until ${suffix}`
+//   } else {
+//     return `${parsedDays} Days Until ${suffix}`
+//   }
+// }
